% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{attr_cfig_parse}
\alias{attr_cfig_parse}
\title{Read and parse the attribute config yaml file to create parameter
list object}
\usage{
attr_cfig_parse(path_attr_config)
}
\arguments{
\item{path_attr_config}{full path to the attribute config file}
}
\description{
Read and parse the attribute config yaml file to create parameter
list object
}
\details{
Parses the attribute config file to generate the parameter
list \code{Retr_Params} used throughout proc.attr.hydfab
}
