% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{build_cfig_path}
\alias{build_cfig_path}
\title{Build a standardized path for every config file used in RaFTS}
\usage{
build_cfig_path(path_known_config, path_or_name_cfig)
}
\arguments{
\item{path_known_config}{The full filepath of a known configuration file}

\item{path_or_name_cfig}{Path or name of configuration file. If only name provided, it's assumed it resides in same directory as \code{path_known_config}}
}
\description{
Build a standardized path for every config file used in RaFTS
}
\details{
Ensure the 'path_known_config' exists
}
\seealso{
\code{rafts_algo.build_cfig_path} The python equivalent of this function
}
