% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{check_attr_selection}
\alias{check_attr_selection}
\title{Check that attributes selected by user are available}
\usage{
check_attr_selection(attr_cfg_path = NULL, vars = NULL, verbose = TRUE)
}
\arguments{
\item{attr_cfg_path}{a path to a .yaml configuration file specifying which
attributes a user is requesting}

\item{vars}{a list specifying which attributes a user is requesting, in lieu
of a list coming from a .yaml configuration file}
}
\description{
Sees if the attributes requested by a user matches with the
attributes supported by the package, listed in the attribute menu.
Returns list of
\itemize{
\item missing_vars: a list of the requested variables that were not found
in the attribute menu as specified.
}
}
\author{
Lauren Bolotin \email{lauren.bolotin@noaa.gov }
}
