% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{check_miss_attrs_comid_io}
\alias{check_miss_attrs_comid_io}
\title{Find comid values that do not have all expected attribute values}
\usage{
check_miss_attrs_comid_io(dt_all, attr_vars, dir_db_attrs)
}
\arguments{
\item{dt_all}{Dataframe/datatable of all locations and attributes}

\item{attr_vars}{List of the data source and expected attributes (e.g. Retr_Params$vars)}

\item{dir_db_attrs}{Directory where attribute data are stored.}
}
\description{
Find comid values that do not have all expected attribute values
}
\details{
Writes to file the missing comid-attribute pairings after
first updating the existing known missing data
}
\seealso{
\link{proc_attr_mlti_wrap}

\link{retr_attr_new}
}
