% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafts_utils.R
\name{compile_chunks_ndplus_geoms}
\alias{compile_chunks_ndplus_geoms}
\title{Compile chunked NHDPlus-retrieved geospatial datasets}
\usage{
compile_chunks_ndplus_geoms(
  dir_save_nhdp,
  seq_nums = NULL,
  filename_str = NULL
)
}
\arguments{
\item{dir_save_nhdp}{The base directory location to store data}

\item{seq_nums}{The numbers expected in each chunked filename}

\item{filename_str}{The custom string for writing data to file}
}
\value{
A compiled list of the geospatial data, containing:
\list{catchment} The catchment boundaries NHDplus dataset
\list{flowlines} The flowlines NHDplus dataset
\list{outlet} The catchment outlet NHDplus dataset
\list{input_dt} The input data.table with data used to query NHDplus
\list{path_gpkg_compiled} File path of the saved geopackage
}
\description{
Compile chunked NHDPlus-retrieved geospatial datasets
}
\seealso{
\link[proc.attr.hydfab]{dl_nhdplus_geoms_wrap}
}
