% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_hydfab.R
\name{custom_hfab_id}
\alias{custom_hfab_id}
\title{Build a custom hydrofabric id that is unique to place}
\usage{
custom_hfab_id(df)
}
\arguments{
\item{df}{The dataframe for a specific location corresponding to the
hydrofabric 'network'}
}
\description{
Build a custom hydrofabric id that is unique to place
}
