% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafts_utils.R
\name{dl_nhdplus_geoms_wrap}
\alias{dl_nhdplus_geoms_wrap}
\title{Grab all nhdplus geometries/comids: outlet, flowlines, catchment}
\usage{
dl_nhdplus_geoms_wrap(
  df,
  col_id,
  dir_save_nhdp,
  filename_str,
  id_type = c("comid", "AOI")[1],
  keep_cols = c(NULL, "all")[1],
  seq_size = 390,
  overwrite_chunk = FALSE
)
}
\arguments{
\item{df}{dataframe with location information that \code{get_nhdplus} uses}

\item{col_id}{The column in the dataframe corresponding to the id}

\item{dir_save_nhdp}{The directory location to store data}

\item{filename_str}{The filename string to use for compiled output data}

\item{id_type}{the type of id corresponding to \code{col_id}. Default 'comid'. 'AOI' may be used for geospatial queries}

\item{keep_cols}{the columns inside \code{df} to also save. May be 'all'. Default NULL means no df data written to file}

\item{seq_size}{the size of the sequence of ids to query within an hour. Recommend <400}

\item{overwrite_chunk}{Should file chunks be overwritten? Default FALSE means skipping database connection & file chunking if a data sequence already created}
}
\description{
Retrieve all nhdplus layers for a comid/AOI, and download/save data
in chunks at hourly intervals to account for external database hits
}
\details{
If AOI selected, the area of interest is assumed to be a line, and
the midpoint is selected for querying NLDI.
Recommended to stick to a certain seq_size (e.g. default) to easily re-use
file chunks when \code{overwrite_chunk} is FALSE.
}
\seealso{
gen_pred_locs_rfcs.R script
}
\author{
Guy Litt
}
