% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_hydfab.R
\name{dms_to_dd}
\alias{dms_to_dd}
\title{Internal function to convert dms to decimal degrees}
\usage{
dms_to_dd(dms)

dms_to_dd(dms)
}
\description{
Internal function to convert dms to decimal degrees

Internal function to convert dms to decimal degrees
}
\seealso{
\link[proc.attr.hydfab]{read_noaa_hads_sites}

\link[proc.attr.hydfab]{read_noaa_hads_sites}
}
