% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{fs_attrs_miss_wrap}
\alias{fs_attrs_miss_wrap}
\title{DEPRECATED. Wrapper searching for comid-attribute data identified as
missing}
\usage{
fs_attrs_miss_wrap(path_attr_config)
}
\arguments{
\item{path_attr_config}{The file path to the attribute config file}
}
\description{
DEPRECATED. Wrapper searching for comid-attribute data identified as
missing
}
\details{
Use fs_attrs_miss_mlti_wrap instead.
Given missing comid-attribute pairings previously identified
from fs_tfrm_attrs.py, and generated as a file by python function
\code{fs_algo.tfrm_attr.write_missing_attrs}
}
\seealso{
\code{fs_algo.tfrm_attr.write_missing_attrs} python

\link{fs_attrs_miss_mlti_wrap}
}
