% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_hydfab.R
\name{get_linestring_midpoint}
\alias{get_linestring_midpoint}
\title{Get the midpoint of a linestring geometry}
\usage{
get_linestring_midpoint(sf_lines)
}
\arguments{
\item{sf_lines}{sfc_LINESTRING class. The geometry of a linestring}
}
\description{
Hydrofabric/NLDI may return linestring objects (e.g. stream
segments). Use this function to extract the single midpoint of that
linestring.
}
\seealso{
\link[proc.attr.hydfab]{retr_hfuids}

\link[proc.attr.hydfab]{retr_hfab_id_wrap}
}
