% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{grab_attrs_datasets_fs_wrap}
\alias{grab_attrs_datasets_fs_wrap}
\title{Grab catchment attributes from processed formulation-selector input}
\usage{
grab_attrs_datasets_fs_wrap(
  Retr_Params,
  lyrs = "network",
  overwrite = FALSE,
  path_save_gpkg_cstm = NULL
)
}
\arguments{
\item{Retr_Params}{list of parameters built for grabbing catchment attribute data. List objects include the following:
\itemize{
\item \code{paths} list of directories or paths used to acquire and save data These include the following:
\item \code{path$dir_db_attrs} local path for saving catchment attributes as parquet files
\item \code{path$paths_ha} the local or s3 location(s) of tabular data file with HydroATLAS downscaled to hydrofabric divide
\item \code{path$dir_std_base} the location of user_data_std containing dataset that were standardized by \pkg{fs_prep}.
\item \code{datasets} character vector. A list of datasets of interest inside \code{paths$dir_std_base}. If 'all' is specified, then all datasets in the directory are processed.
\item \code{ds_type} character. The identifier to use in filename when writing attribute location metadata retrieved from nhdplusTools::get_nldi_feature()
}}

\item{lyrs}{default "network" the hydrofabric layers of interest.
Only 'network' is needed for attribute grabbing.}

\item{overwrite}{boolean default FALSE. Should hydrofabric data be overwritten?}

\item{path_save_gpkg_cstm}{A custom path to save the geopackage mapping
comid/gage_id/coords. Default NULL means that the path is automated based
on the standardized input dataset location \link[proc.attr.hydfab]{std_path_retr_gpkg}}
}
\description{
Wrapper function that acquires catchment attribute data from
formulation-selector processed input generated via \pkg{fs_prep} package
Returns list of data.table object of nhdplusTools::get_nldi_feature()
for all gage_ids
}
\details{
Runs two proc.attr.hydfab functions:
\link[proc.attr.hydfab]{proc_attr_read_gage_ids_fs} - retrieves the gage_ids generated by \pkg{fs_prep}
\link[proc.attr.hydfab]{proc_attr_gageids} - retrieves the attributes for all provided gage_ids
}
