% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{hfab_config_opt}
\alias{hfab_config_opt}
\title{Configure hydrofabric-relevant optional params}
\usage{
hfab_config_opt(
  hfab_config,
  reqd_hfab = c("s3_base", "s3_bucket", "hf_cat_sel", "source")
)
}
\arguments{
\item{hfab_config}{The hydrofabric-specific section from the config file, hydfab_config. list.}

\item{reqd_hfab}{The non-optional item names in the hydrofabric config file}
}
\value{
List with default arguments populated corresponding to hfsubetR::get_subset(),
}
\description{
If an argument provided in the config file is NULL, first look
for default param value from the \code{proc.attr_hydfab::proc_attr_hf}.
If that is NULL, then look for default value from the
\code{hfsubsetR::get_subset()} args if that param exists there. Otherwise,
uses default param value in \code{proc.attr_hydfab::proc_attr_wrap}.
}
