% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{io_attr_dat}
\alias{io_attr_dat}
\title{Write the updated basin attribute data.table}
\usage{
io_attr_dat(
  dt_new_dat,
  path_attrs,
  distinct_cols = c("featureID", "data_source", "attribute")
)
}
\arguments{
\item{path_attrs}{parquet filepath for attribute data}

\item{distinct_cols}{The column names in dt_new_dat that must be distinct}

\item{dt_cmbo}{The standardized data.table of attributes}
}
\description{
Write the updated basin attribute data.table
}
\details{
Checks to see if data already exists. If so, read it in. Then
merges new data with existing data and remove any duplicates
}
\seealso{
\link{retrieve_attr_exst} for retrieving existing attributes

\link{std_attr_data_fmt}

\link{std_path_attrs}
}
