% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{map_attrs_to_dataset}
\alias{map_attrs_to_dataset}
\title{Match the variable names to the data source}
\usage{
map_attrs_to_dataset(vars)
}
\arguments{
\item{vars}{List of variable names, possibly mixed across different data sources}
}
\description{
Match the variable names to the data source
}
