% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_hydfab.R
\name{map_hfab_oconus_sources_wrap}
\alias{map_hfab_oconus_sources_wrap}
\title{Match the hydrofabric OCONUS geopackage locations to missing hydrofabric data}
\usage{
map_hfab_oconus_sources_wrap(
  dt_need_hf,
  hfab_srce_map,
  col_lat = "latitude",
  col_lon = "longitude"
)
}
\arguments{
\item{dt_need_hf}{data.table of locations that need a hydrofabric id. Must include lat/lon cols}

\item{hfab_srce_map}{OCONUS postal code ids (colname 'domain') mapped to paths to respective gpkg files}

\item{col_lat}{The latitude column name inside \code{dt_need_hf}}

\item{col_lon}{The longitude column name inside \code{dt_need_hf}}
}
\description{
Match the hydrofabric OCONUS geopackage locations to missing hydrofabric data
}
\seealso{
\link[proc.attr.hydfab]{retr_state_terr_postal}
}
