% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{parse_attr_srce_type_config}
\alias{parse_attr_srce_type_config}
\title{Parse the attr_source_types.yml config file internal to proc.attr.hydfab}
\usage{
parse_attr_srce_type_config()
}
\description{
Parse the attr_source_types.yml config file internal to proc.attr.hydfab
}
\seealso{
\link[proc.attr.hydfab]{map_attrs_to_dataset}
}
