% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_hydatl_hydfab.R
\name{plot_hf_ha}
\alias{plot_hf_ha}
\title{Plot hydrofabric and HydroATLAS basins}
\usage{
plot_hf_ha(
  div_hf_hab,
  viz_idx = 1,
  total_vpu = FALSE,
  col_geom_ha = "geometry",
  col_geom_hf = "geom",
  col_centroid_hf = "centroid"
)
}
\arguments{
\item{div_hf_hab}{merged dataframe of hydrofabric and HydroATLAS}

\item{viz_idx}{Row indices of dataframe for plotting. Default 1.}

\item{total_vpu}{Boolean. Should the entire dataframe's gemoetry also be included?}
}
\description{
Plot hydrofabric and HydroATLAS basins
}
\details{
Will calculate hydrofabric centroids if column doesn't exist
}
