% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{proc_attr_exst_wrap}
\alias{proc_attr_exst_wrap}
\title{Existing attribute data checker}
\usage{
proc_attr_exst_wrap(comid, path_attrs, vars_ls, bucket_conn = NA)
}
\arguments{
\item{comid}{character class. The common identifier USGS location code for a surface water feature.}

\item{path_attrs}{character. Path to attribute file data storage location}

\item{vars_ls}{list. Variable names}

\item{bucket_conn}{TODO add cloud conn details in case data stored in s3}
}
\description{
Retrieves what attribute data already exists in a data storage
path for a given comid and identifies missing attributes.
Returns list of
\itemize{
\item dt_all: a data.table of existing comid data,
\item need_vars: a list of datasource ids containing a list of variable
names that will be downloaded.
}
}
\seealso{
\code{\link[=proc_attr_wrap]{proc_attr_wrap()}}
}
\author{
Guy Litt \email{guy.litt@noaa.gov}
}
