% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{proc_attr_hf}
\alias{proc_attr_hf}
\title{Retrieve hydrofabric data of interest based on location identifier}
\usage{
proc_attr_hf(
  comid,
  dir_db_hydfab,
  custom_name = "{lyrs}_",
  fileext = "gpkg",
  lyrs = c("divides", "network")[2],
  hf_cat_sel = TRUE,
  overwrite = FALSE,
  hf_version = NULL,
  type = NULL,
  domain = NULL
)
}
\arguments{
\item{comid}{character class. The common identifier USGS location code for a surface water feature.}

\item{dir_db_hydfab}{character class. Local directory path for storing hydrofabric data}

\item{custom_name}{character class. A custom name to insert into hydrofabric file. Default \code{glue("{lyrs}_")}}

\item{fileext}{character class. file extension of hydrofabric file. Default 'gpkg'}

\item{lyrs}{character class. The layer name(s) of interest from hydrofabric. Default 'network'.}

\item{hf_cat_sel}{boolean. TRUE for a total catchment characterization specific to a single comid, FALSE (or anything else) for all subcatchments}

\item{overwrite}{boolean. Overwrite local data when pulling from hydrofabric s3 bucket? Default to FALSE.}

\item{hf_version}{character class. The hydrofabric version. When NULL, defaults to same as \code{hfsubsetR::get_subset()}}

\item{type}{hydrofabric type. When NULL, defaults to same as \code{hfsubsetR::get_subset()}, likely 'nextgen'}

\item{domain}{hydrofabric domain. When NULL, defaults to same as \code{hfsubsetR::get_subset()}, likely 'conus'}
}
\description{
Checks to see if a local dataset exists. If not, retrieve from lynker-spatial s3 bucket
}
\author{
Guy Litt \email{guy.litt@noaa.gov}
}
