% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{proc_attr_hydatl}
\alias{proc_attr_hydatl}
\title{Retrieve hydroatlas variables}
\usage{
proc_attr_hydatl(hf_id, s3_path, ha_vars, local_path = NA)
}
\arguments{
\item{hf_id}{numeric. the hydrofabric id, expected to be the COMID}

\item{s3_path}{character. full path to the s3 bucket's file holding the hydroatlas data}

\item{ha_vars}{list of characters. The variables of interest in the hydroatlas v1}

\item{local_path}{character. The local filepath where hydroatlas data are saved to reduce s3 bucket connections.}
}
\description{
retrieves hydrofabric variables from s3 bucket
}
