% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{proc_attr_hydatl_wrap}
\alias{proc_attr_hydatl_wrap}
\title{Process hydroatlas attributes wrapper}
\usage{
proc_attr_hydatl_wrap(
  hf_ids,
  paths_ha,
  ha_vars,
  hf_id_cols = c("hf_uid", "hf_id", "id")
)
}
\arguments{
\item{hf_id_cols}{Possible hydrofabric id columns, \emph{listed in priority}, meaning
the \code{'hf_uid'} is the most important column to find, and if that is not
present, move on to the \code{'hf_id'} and finally to \code{'id'}}
}
\description{
Process hydroatlas attributes wrapper
}
\details{
The hf_ids may contain 1) a COMID, corresponding to CONUS, &/or
2) a hydrofabric id for OCONUS
}
