% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{proc_attr_read_gage_ids_fsds}
\alias{proc_attr_read_gage_ids_fsds}
\title{Read in standardized FSDS gage_id location identifiers}
\usage{
proc_attr_read_gage_ids_fsds(dir_dataset, ds_filenames = "")
}
\arguments{
\item{dir_dataset}{directory path to the dataset}

\item{ds_filenames}{a matching string specific to dataset(s) of interest
inside \code{dir_dataset}}
}
\description{
Reads output generated using \pkg{fsds_proc} python package and
selects the gage_id location identifier and the attributes of interest
}
\details{
Returns a list of the following objects:
gage_ids: array of gage_id values
featureSource: The type of nhdplus feature source corresponding to gage_id
featureID: The method of converting gage_id into a standardized featureSource's featureID
}
\seealso{
\link{[get_nldi_features]}
}
