% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{proc_attr_std_hfsub_name}
\alias{proc_attr_std_hfsub_name}
\title{Standardidze hydrofabric subsetter's local filename}
\usage{
proc_attr_std_hfsub_name(comid, custom_name = "", fileext = "gpkg")
}
\arguments{
\item{comid}{the USGS common identifier, generated by nhdplusTools}

\item{custom_name}{Desired custom name following 'hydrofab_'}

\item{fileext}{file extension of the hydrofrabric data. Default 'gpkg'}
}
\description{
Internal function that ensures consistent filename
}
