% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{proc_attr_wrap}
\alias{proc_attr_wrap}
\title{Wrapper to retrieve variables when processing attributes}
\usage{
proc_attr_wrap(comid, Retr_Params, lyrs = "network", overwrite = FALSE)
}
\description{
Wrapper to retrieve variables when processing attributes
}
