% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{rafts_retr_nhdp_comids_geom_wrap}
\alias{rafts_retr_nhdp_comids_geom_wrap}
\title{Read or generate a sf object from NHDplus queries for comid, and
.  update file with any newly retrieved locations}
\usage{
rafts_retr_nhdp_comids_geom_wrap(
  path_save_gpkg,
  gage_ids,
  featureSource = "nwissite",
  featureID = "USGS-{gage_id}",
  epsg = 4326
)
}
\arguments{
\item{path_save_gpkg}{The filepath where the geopackage file should live}

\item{gage_ids}{array of gage_id values to be queried for catchment attributes}

\item{featureSource}{The \link[nhdplusTools]{get_nldi_feature}feature featureSource,
e.g. 'nwissite'}

\item{featureID}{a glue-configured conversion of gage_id into a recognized
featureID for  \link[nhdplusTools]{get_nldi_feature}. E.g. if gage_id
represents exactly what the nldi_feature$featureID should be, then
featureID="{gage_id}". In other instances, conversions may be necessary,
e.g. featureID="USGS-{gage_id}". When defining featureID, it's expected
that the term 'gage_id' is used as a variable in glue syntax to create featureID}

\item{epsg}{The EPSG code to use for the CRS; nhdplus default is 4326}
}
\description{
Try reading geopackage file, if it doesn't exist or is missing
locations based on the gage_ids of interest, grab them & update the
geopackage file
}
\seealso{
\link[proc.attr.hydfab]{proc_attr_read_gage_ids_fs}

\link[proc.attr.hydfab]{rafts_retr_nhdp_comids_geom}

\code{rafts_algo.utils.rafts_retr_nhdp_comids_geom_wrap}
}
