% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{read_attr_srce_type_config}
\alias{read_attr_srce_type_config}
\title{Read the attribute source type config file}
\usage{
read_attr_srce_type_config()
}
\description{
Read the attribute source type config file
}
