% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{read_fs_attr_menu_config}
\alias{read_fs_attr_menu_config}
\title{Read in proc.attr.hydfab package's extdata describing attributes & data sources}
\usage{
read_fs_attr_menu_config()
}
\description{
Read in proc.attr.hydfab package's extdata describing attributes & data sources
}
