% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{read_fs_retr_gpkg}
\alias{read_fs_retr_gpkg}
\title{Read geopackage containing identifers and point locations}
\usage{
read_fs_retr_gpkg(path_save_gpkg, verbose = FALSE)
}
\arguments{
\item{path_save_gpkg}{filepath to the geopackage}
}
\description{
Reads in a geopackage whose filepath is
standardized by \link[proc.attr.hydfab]{std_path_retr_gpkg}
and created by \link[proc.attr.hydfab]{fs_retr_nhdp_comids_geom_wrap}
}
\details{
The renaming may be a temporary solution just-in case file
originated in python \code{fs_algo}
}
\seealso{
\link[proc.attr.hydfab]{fs_retr_nhdp_comids_geom_wrap}

\link[proc.attr.hydfab]{std_path_retr_gpkg}

\link[proc.attr.hydfab]{fs_retr_nhdp_comids_geom_wrap}
}
