% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_hydfab.R
\name{read_hfab_layer}
\alias{read_hfab_layer}
\title{Read a hydrofabric layer}
\usage{
read_hfab_layer(path_gpkg, layer)
}
\arguments{
\item{path_gpkg}{The filepath of the geopackage for a hydrofabric domain}

\item{layer}{The layer of interest. Options include
'divides','flowpaths','network','nexus'}
}
\description{
Read a hydrofabric layer
}
\details{
Performs a vpu value correction in case of AK labeled as 'hi'.
Designed during hydrofabric v2.2
}
