% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_hydfab.R
\name{read_hfab_layers}
\alias{read_hfab_layers}
\title{Read the hydrofabric file gpkg for each layer}
\usage{
read_hfab_layers(path_gpkg, layers = NULL)
}
\arguments{
\item{path_gpkg}{The filepath of the geopackage for a hydrofabric domain}

\item{layers}{The hydrofabric layers of interest. Default NULL means all.
Other options include any of the following inside
\code{c("divides", "flowpaths", "network", "nexus")}}
}
\description{
Read the hydrofabric file gpkg for each layer
}
\seealso{
proc_attr_hf
}
