% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_hydatl_hydfab.R
\name{read_hfab_lyr_val}
\alias{read_hfab_lyr_val}
\title{Read the hydrofabric layer, convert to WGS 84, and make valid}
\usage{
read_hfab_lyr_val(path_hfab, lyr = "divides")
}
\arguments{
\item{path_hfab}{filepath to the hydrofabric geopackage of interest}

\item{lyr}{The layer to read from the geopackage, default \code{'divides'}}
}
\description{
Read the hydrofabric layer, convert to WGS 84, and make valid
}
