% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_hydatl_hydfab.R
\name{read_hydatl_by_vpu_val}
\alias{read_hydatl_by_vpu_val}
\title{Read the HydroATLAS by vpu value}
\usage{
read_hydatl_by_vpu_val(vpu, dir_base_hydatl)
}
\arguments{
\item{vpu}{The OCONUS VPU id used in the hydrofabric. Options}
}
\description{
Define Paths to HydroATLAS data downloaded from HydroATLAS website
based on the hydrofabric vpu identifiers. Make geometries valid.
}
