% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{read_loc_data}
\alias{read_loc_data}
\title{Read location identifiers}
\usage{
read_loc_data(loc_id_filepath, loc_id, fmt = "csv")
}
\arguments{
\item{loc_id_filepath}{csv filepath or dataset filepath/directory.}

\item{loc_id}{The column name of the identifier column}

\item{fmt}{The format passed to arrow::open_dataset() in the non-csv case.
Default 'csv'. May also be 'parquet', 'arrow', 'feather', 'zarr', etc.}
}
\description{
Reads directly from a csv or arrow-compatible dataset.
Returns the dataset's column identifer renamed as 'gage_id' in a tibble
}
\seealso{
\code{\link[=proc_attr_read_gage_ids_fsds]{proc_attr_read_gage_ids_fsds()}}

\code{\link[=proc_attr_wrap]{proc_attr_wrap()}}
}
