% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_hydfab.R
\name{read_noaa_hads_sites}
\alias{read_noaa_hads_sites}
\title{Read and parse NOAA HADS sites metadata}
\usage{
read_noaa_hads_sites(
  path_hads_sites = "https://hads.ncep.noaa.gov/USGS/ALL_USGS-HADS_SITES.txt",
  name_map = data.frame(hads_row2 = c("5 CHR", "STATION", "GOES", "NWS", "LATITUDE",
    "LONGITUDE", "LOCATION"), std_name = c("lid", "usgsId", "GOES", "NWS_HSA",
    "latitude", "longitude", "name"))
)
}
\arguments{
\item{path_hads_sites}{The path to NCEP NOAA HADS site locations}

\item{name_map}{Mapping the default hads column names to desired names
consistent with those used by NWPS in the nwps_all_gauges_report.csv and
more importantly, the NWPS data retrieved via retr_noaa_gauges_meta}
}
\description{
Read and parse NOAA HADS sites metadata
}
\seealso{
\link[proc.attr.hydfab]{retr_noaa_gauges_meta}
}
