% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_hydfab.R
\name{read_noaa_nwps_gauges}
\alias{read_noaa_nwps_gauges}
\title{Read & clean NOAA NWPS gauge metadata}
\usage{
read_noaa_nwps_gauges(
  path_nwps_rpt =
    "https://water.noaa.gov/resources/downloads/reports/nwps_all_gauges_report.csv"
)
}
\arguments{
\item{path_nwps_rpt}{Path to the data source}
}
\description{
Read & clean NOAA NWPS gauge metadata
}
