% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_hydfab.R
\name{retr_hf_id_xy}
\alias{retr_hf_id_xy}
\title{Retrieve the hydrofabric id using xy coordinates query to hydrofabric gpkg}
\usage{
retr_hf_id_xy(df_sf, path_gpkg, geom_col = "geometry")
}
\arguments{
\item{df_sf}{data.frame with a POINT geometry column}

\item{geom_col}{The column name for the sf geometry object in df_sf. Default 'geometry'}

\item{path_hf}{filepath to the hydrofabric gpkg of interest}
}
\description{
Given an sf object with a POINT geometry column \code{geom_col},
pass the point column name, search for the location using hfsubsetR,
and pick the nearest downstream point with a hydrofabric id.
}
\details{
For CONUS locations, the comid will be returned.
}
\seealso{
\link[proc.attr.hydfab]{retr_hfab_id_coords}

\link[proc.attr.hydfab]{retr_comids_coords}
}
