% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_hydfab.R
\name{retr_hfab_id_usgs_gage}
\alias{retr_hfab_id_usgs_gage}
\title{NEEDS MORE WORK Retrieve the hydrofabric ID from the network via the USGS gage_id}
\usage{
retr_hfab_id_usgs_gage(gage_id, ntwk)
}
\arguments{
\item{gage_id}{The USGS identifier}

\item{ntwk}{data.frame network layer from the appropriate hydrofabric}
}
\description{
NEEDS MORE WORK Retrieve the hydrofabric ID from the network via the USGS gage_id
}
\details{
Must provide the appropriate network! E.g. specific to domain,
such as Alaska
}
