% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_hydfab.R
\name{retr_hfab_id_wrap}
\alias{retr_hfab_id_wrap}
\title{Retrieve hydrofabric IDs wrapper}
\usage{
retr_hfab_id_wrap(
  dt_need_hf,
  path_oconus_hfab_config,
  col_usgsId = "usgsId",
  col_lon = "longitude",
  col_lat = "latitude",
  epsg_coords = 4326
)
}
\arguments{
\item{dt_need_hf}{data.table of needed locations}

\item{path_oconus_hfab_config}{File path to yaml config mapping of OCONUS
hydrofabric locations, postal codes, and expected CRS geopackage filepaths}

\item{col_usgsId}{column name inside \code{dt_need_hf} for the USGS gage ID}

\item{col_lon}{column name inside \code{dt_need_hf} for longitude value}

\item{col_lat}{column name inside \code{dt_need_hf} for latitude value}

\item{epsg_coords}{The CRS for the lat/lon data inside \code{dt_need_hf}}
}
\description{
Retrieve the unique ids used in the hydrofabric, as an hf_uid.
If an ID isn't unique (e.g. an oCONUS hf_id as of hydrofabric v2.2,
then a unique id is built)
}
\details{
Intended for situations when comids unavailable, generally as OCONUS
More work needed to make this fully compatible with CONUS locations.
}
\seealso{
prep_oconus_hydroatlas.R also creates the hf_uid using custom_hf_id()
}
