% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_hydfab.R
\name{retr_hfuids}
\alias{retr_hfuids}
\title{Retrieve the hydrofabric unique id}
\usage{
retr_hfuids(
  loc_ids,
  path_oconus_hfab_config,
  featureSource = c("nwissite", "wqp", "comid", "location")[1]
)
}
\arguments{
\item{loc_ids}{The location identifier whose form is described by
\code{featureSource}}

\item{path_oconus_hfab_config}{Path to the config file defining where
the hydrofabric files are stored.}

\item{featureSource}{The \link[nhdplusTools]{get_nldi_feature}
featureSource, which is attempted to be translated into the
\link[dataRetrieval]{findNLDI} featureSource types. Default 'nwissite'
means the argument \code{nwis} is used in \code{findNLDI}, which is likely the
best option in the case of USGS gage ids. Other options include 'wqp',
for the water quality portal (fewer locations means less data returned),
'comid', and 'location'. Location values in \code{loc_ids}
would need to be defined as two points (e.g. \code{c(-115,40)})}
}
\description{
Intended for oCONUS locations
\link[proc.attr.hydfab]{retr_comids} can grab the comids from CONUS
locations. Thus this function is intended to fill the gaps after
\link[proc.attr.hydfab]{retr_comids} has run.
}
\seealso{
\link[proc.attr.hydfab]{retr_comids} Companion function for retrieiving comids across CONUS

\link[proc.attr.hydfab]{retr_hfab_id_wrap} Specifically designed for retrieving the hydrofabric unique id, intended for oCONUS

\link[proc.attr.hydfab]{proc_attr_gage_ids}
}
