% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_hydfab.R
\name{retr_noaa_gauges_meta}
\alias{retr_noaa_gauges_meta}
\title{Retrieve metadata based on a NOAA RFC gauge ID, aka lid}
\usage{
retr_noaa_gauges_meta(
  gauge_ids,
  gauge_url_base = "https://api.water.noaa.gov/nwps/v1/gauges",
  retr_ids = c("lid", "usgsId", "name", "latitude", "longitude")
)
}
\arguments{
\item{gauge_ids}{list of NOAA gauge ids of interest}

\item{gauge_url_base}{the base api url for NWPS}

\item{retr_ids}{The desired data to retrieve from the api}
}
\description{
Uses the NWPS api to retrieve gauge metadata
}
\seealso{
\link[proc.attr.hydfab]{read_noaa_hads_sites}
}
