% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_hydfab.R
\name{retr_state_terr_postal}
\alias{retr_state_terr_postal}
\title{Retrieve postal id of state or territory}
\usage{
retr_state_terr_postal(lat, lon)
}
\arguments{
\item{lat}{latitude in epsg 4326}

\item{lon}{longitude in epsg 4326}
}
\description{
Retrieve postal id of state or territory
}
