% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R, R/rafts_utils.R
\name{retrieve_attr_exst}
\alias{retrieve_attr_exst}
\title{Grab previously-aggregated attributes from locations of interest}
\usage{
retrieve_attr_exst(comids, vars, dir_db_attrs, bucket_conn = NA)

retrieve_attr_exst(comids, vars, dir_db_attrs, bucket_conn = NA)
}
\arguments{
\item{comids}{character class. The comids/hf_uids/etc. of interest.}

\item{vars}{character class. The attribute variables of interest.}

\item{dir_db_attrs}{character class. The path where data}

\item{bucket_conn}{Default NA. Placeholder in case a bucket connection is
ever created}
}
\description{
Retrieves existing attribute data already stored in the
dir_db_attrs directory as .parquet files & return tbl of all comids and
attributes of interest.

Retrieves existing attribute data already stored in the
dir_db_attrs directory as .parquet files & return tbl of all comids and
attributes of interest.
}
\details{
Only considers data already generated inside dir_db_attrs. If
more data are needed, acquire attribute data acquisition using
\link[proc.attr.hydfab]{proc_attr_wrap}.
Runs checks on input arguments and retrieved contents, generating warnings
if requested comids and/or variables were completely absent from the dataset

Only considers data already generated inside dir_db_attrs. If
more data are needed, acquire attribute data acquisition using proc_attr_wrap().
Runs checks on input arguments and retrieved contents, generating warnings
if requested comids and/or variables were completely absent from the dataset
}
\seealso{
\link[proc.attr.hydfab]{proc_attr_wrap}

\link[proc.attr.hydfab]{proc_attr_wrap}
}
