% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{std_dir_dataset}
\alias{std_dir_dataset}
\title{Generate the standardized dataset directory}
\usage{
std_dir_dataset(dir_std_base, ds)
}
\arguments{
\item{dir_std_base}{The standardized base directory}

\item{ds}{The dataset name (aka dir name)}
}
\description{
Generate the standardized dataset directory
}
\seealso{
\link[proc.attr.hydfab]{std_path_retr_gpkg}
}
