% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafts_utils.R
\name{std_dir_gpkg_chunk_nhdp_geom}
\alias{std_dir_gpkg_chunk_nhdp_geom}
\title{Generate the chunk subdirectory for temporarily storing results}
\usage{
std_dir_gpkg_chunk_nhdp_geom(dir_save_nhdp)
}
\arguments{
\item{dir_save_nhdp}{The base directory location containing the chunk dir}
}
\description{
Recursively creates the chunk directory if it doesn't exist
}
\seealso{
\link[proc.attr.hydfab]{dl_nhdplus_geoms_wrap}
}
