% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_hydfab.R
\name{std_feat_id}
\alias{std_feat_id}
\title{Generate the standardized format of featureID and featureSource}
\usage{
std_feat_id(
  df,
  name_featureSource = c("COMID", "custom_hfuid")[1],
  col_featureID = NULL,
  vals_featureID = NULL
)
}
\arguments{
\item{df}{The data.frame of interest containing an identifier}

\item{name_featureSource}{The expected name of the feature source. Default "COMID" expected for CONUS, but OCONUS will use "custom_hfuid"}

\item{col_featureID}{The column name inside \code{df} containing featureID of interest. Must provide if \code{vals_featureID} empty.}

\item{vals_featureID}{The unique identifiers to populate featureID of interest. Must provide if \code{col_featureID} empty.}
}
\description{
Generate the standardized format of featureID and featureSource
}
