% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{std_miss_path}
\alias{std_miss_path}
\title{standardize path to file listing all missing attributes}
\usage{
std_miss_path(dir_db_attrs)
}
\arguments{
\item{dir_db_attrs}{The directory to the attribute database}
}
\description{
standardize path to file listing all missing attributes
}
\seealso{
\code{fs_algo.tfrm_attrs.std_miss_path} python package
}
