% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_hydatl_hydfab.R
\name{std_path_attrs_all_parq}
\alias{std_path_attrs_all_parq}
\title{standardize path to they hydroatlas attributes for all OCONUS
hydrofabric domains}
\usage{
std_path_attrs_all_parq(dir_base_hfab, ls_vpus = c("ak", "prvi"))
}
\arguments{
\item{dir_base_hfab}{Base directory where hydrofabric data stored}

\item{ls_vpu}{the hydrofabric vpus of interest (e.g. \code{c('ak','prvi')})}
}
\description{
standardize path to they hydroatlas attributes for all OCONUS
hydrofabric domains
}
\details{
Intended for just ak and prvi
}
