% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{std_path_dataset}
\alias{std_path_dataset}
\title{Return filepath of standardized rafts gage_id
location identifiers}
\usage{
std_path_dataset(dir_dataset, ds_filenames = "")
}
\arguments{
\item{dir_dataset}{directory path containing the dataset}

\item{ds_filenames}{a matching string specific to dataset(s) of interest
inside \code{dir_dataset}}
}
\description{
Return filepath of standardized rafts gage_id
location identifiers
}
