% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafts_utils.R
\name{std_path_gpkg_all_nhdp_geom}
\alias{std_path_gpkg_all_nhdp_geom}
\title{the standardized path for saving all geopackages}
\usage{
std_path_gpkg_all_nhdp_geom(dir_save_nhdp, filename_str)
}
\arguments{
\item{dir_save_nhdp}{The directory location to store data}

\item{filename_str}{The filename string to use for compiled output data}
}
\description{
the standardized path for saving all geopackages
}
\seealso{
\link[proc.attr.hydfab]{dl_nhdplus_geoms_wrap}
}
