% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafts_utils.R
\name{std_path_gpkg_chunk_nhdp_geom}
\alias{std_path_gpkg_chunk_nhdp_geom}
\title{standardize each chunked filepath}
\usage{
std_path_gpkg_chunk_nhdp_geom(dir_save_nhdp, seq_num)
}
\arguments{
\item{dir_save_nhdp}{The base directory location containing the chunk dir}

\item{seq_num}{The unique number defining the chunk sequence (expected to
integer from 1 to n number of chunked files)}
}
\description{
standardize each chunked filepath
}
\seealso{
\link[proc.attr.hydfab]{dl_nhdplus_geoms_wrap}
}
