% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{std_path_gpkg_db}
\alias{std_path_gpkg_db}
\title{The filepath to the large gpkg database containing all locations
ever acquired when running RaFTS}
\usage{
std_path_gpkg_db(dir_db_gpkg)
}
\arguments{
\item{dir_db_gpkg}{The directory of the geopackage database}
}
\description{
The filepath to the large gpkg database containing all locations
ever acquired when running RaFTS
}
\seealso{
\link[proc.attr.hydfab]{attr_cfig_parse} creates the directory
structure for \code{dir_db_gpkg}
}
