% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafts_utils.R
\name{std_path_log}
\alias{std_path_log}
\title{Define the path for storing log file specific to an atttribute config}
\usage{
std_path_log(dir_log, path_attr_config, script = "")
}
\arguments{
\item{dir_log}{The RaFTS project's standard log directory}

\item{path_attr_config}{Filepath to the attribute config file}

\item{script}{The string of the script name to add to the log's filename. Default ''.}
}
\description{
Define the path for storing log file specific to an atttribute config
}
