% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{std_path_map_loc_ids}
\alias{std_path_map_loc_ids}
\title{Standardize the path of the csv file that maps NLDI IDs to comids}
\usage{
std_path_map_loc_ids(dir_db_attrs)
}
\arguments{
\item{dir_db_attrs}{The attributes database path}
}
\description{
Uses a sub-directory in the dir_db_attrs to place data
}
