% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{std_path_miss_tfrm}
\alias{std_path_miss_tfrm}
\title{Create a standardized path for storing missing comid-attribute
pairings needed for attribute transformation}
\usage{
std_path_miss_tfrm(dir_db_attrs)
}
\arguments{
\item{dir_db_attrs}{The directory to the attribute database}
}
\description{
Create a standardized path for storing missing comid-attribute
pairings needed for attribute transformation
}
\seealso{
\code{rafts_algo.tfrm_attrs.std_path_miss_tfrm} python package
}
