% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafts_utils.R
\name{std_path_rds_chunk_nhdp_geom}
\alias{std_path_rds_chunk_nhdp_geom}
\title{the standardized filepath for chunked RDS files}
\usage{
std_path_rds_chunk_nhdp_geom(path_gpkg_chunk)
}
\arguments{
\item{path_gpkg_chunk}{The full filepath to a chunked .gpkg}
}
\description{
the standardized filepath for chunked RDS files
}
\seealso{
\link[proc.attr.hydfab]{dl_nhdplus_geoms_wrap}

\link[proc.attr.hydfab]{std_path_gpkg_chunk_nhdp_geom}
}
