% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{std_path_retr_gpkg_wrap}
\alias{std_path_retr_gpkg_wrap}
\title{Wrapper to create the standardized geopackage path for coordinate
. data and id mapping corresponding to the standardized input data from
the \code{rafts_prep} python processing steps}
\usage{
std_path_retr_gpkg_wrap(dir_std_base, ds)
}
\arguments{
\item{dir_std_base}{The standardized base directory}

\item{ds}{The dataset name (folder inside \code{dir_std_base})}
}
\description{
Given the standardized base directory and the dataset name,
generate the standardized geopackage
}
\details{
These same steps happen inside \code{grab_attrs_datasets_rafts_wrap}
}
