% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_hydatl_hydfab.R
\name{std_paths_attrs}
\alias{std_paths_attrs}
\title{standardize path to the HydroATLAS attributes corresponding to
each hydrofabric divide, saved as a csv}
\usage{
std_paths_attrs(dir_base_hfab, vpu)
}
\arguments{
\item{dir_base_hfab}{Base directory where hydrofabric data stored}

\item{vpu}{the hydrofabric vpu of interest (e.g. \code{'ak'} or \code{'prvi'})}
}
\description{
standardize path to the HydroATLAS attributes corresponding to
each hydrofabric divide, saved as a csv
}
