% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_hydatl_hydfab.R
\name{std_paths_ntrsct}
\alias{std_paths_ntrsct}
\title{Standardizing the paths to hydrofabric-HydroATLAS intersection data}
\usage{
std_paths_ntrsct(path_hfab)
}
\description{
Standardizing the paths to hydrofabric-HydroATLAS intersection data
}
