% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafts_utils.R
\name{std_write_gpkg_all_nhdp_geom}
\alias{std_write_gpkg_all_nhdp_geom}
\title{Write standardized .gpkg file, enforcing expected layer names}
\usage{
std_write_gpkg_all_nhdp_geom(
  df_obj,
  path_save_gpkg_all,
  layer_save = c("catchment", "flowlines", "outlet", "input_df")[4]
)
}
\arguments{
\item{df_obj}{The dataframe/sf nhdplus geometry or input data}

\item{path_save_gpkg_all}{The gpkg filepath for writing}

\item{layer_save}{The standardized names for saving different types of
layers to the geopackage file.}
}
\description{
Write standardized .gpkg file, enforcing expected layer names
}
\seealso{
\link[proc.attr.hydfab]{dl_nhdplus_geoms_wrap}
}
