% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafts_utils.R
\name{sub_hf_huc_conus}
\alias{sub_hf_huc_conus}
\title{Grab approximate hydrofabric polygons for HUC levels of interest}
\usage{
sub_hf_huc_conus(huc_level = 8, hydrofab_path = NULL, dir_save_hf_sub)
}
\arguments{
\item{huc_level}{integer. The HUC level (6, 8, 10, 12) to which the user
wants to subset and aggregate hydrofabric catchments. Larger HUCs (e.g. 2,
4) are not compatible with this methodology.}

\item{hydrofab_path}{The local path to the nextgen CONUS hydrofabric gpkg.
Only hydrofabric version 2.2 is currently supported.}
}
\description{
Identifies plausible outlets for catchments at the specified
HUC level by identifying the most downstream reach according to the hydro-
fabric, subsetting the network above this reach, union-ing all divides into
one polygon, and appends to a dataframe, which gets saved as a geopackage.
}
\details{
The outlet locations used for subsetting have not been quality
checked -- users should evaluate if the resulting polygons meet their
needs and criteria.
}
\seealso{
\link[proc.attr.hydfab]{proc_attr_wrap}
}
