% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{tform_cfig_parse}
\alias{tform_cfig_parse}
\title{Read & parse the transformation config yaml file}
\usage{
tform_cfig_parse(path_tfrm_config)
}
\arguments{
\item{path_tfrm_config}{Full filepath to the transformation config file}
}
\description{
Read & parse the transformation config yaml file
}
\seealso{
\link[proc.attr.hydfab]{attr_cfig_parse}
}
