% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{uniq_id_loc_attr}
\alias{uniq_id_loc_attr}
\title{define the unique identifier of comid-attribute pairings}
\usage{
uniq_id_loc_attr(comids, attrs)
}
\description{
define the unique identifier of comid-attribute pairings
}
\seealso{
\link[proc.attr.hydfab]{rafts_attrs_miss_mlti_wrap}
}
