% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{wrap_check_vars}
\alias{wrap_check_vars}
\title{Internal wrapper to run checks on requested attribute variable names}
\usage{
wrap_check_vars(vars_ls)
}
\arguments{
\item{vars_ls}{A named list from Retr_Params$vars in the standardized format}
}
\description{
Given a list of variable categories, each containing vectors
of variable names, check the following:
\enumerate{
\item the variable category is a recognized category name (e.g. 'usgs_vars')
\item the variable names inside the category name are actual variable names
that can be used to retrieve attributes (e.g. 'TOT_TWI' as an nhdplus attribute)
}
}
