% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{write_meta_nldi_feat}
\alias{write_meta_nldi_feat}
\title{Write metadata from NLDI retrieval}
\usage{
write_meta_nldi_feat(dt_site_feat, path_meta)
}
\arguments{
\item{dt_site_feat}{data.table or data.frame of NLDI site features
retrieved using nhdplusTools::get_nldi_feature() and organized
using proc.attr.hydfab::proc_attr_gageids}

\item{path_meta}{the filepath for writing nldi metadata. May be parquet or csv file.}
}
\description{
Standard approach for saving metadata for each location corresponding to
a dataset. The metadata file contains the featureID/featureSource,
data_source, dl_timestamp, gage_id, attribute and its value
}
\seealso{
\link[proc.attr.hydfab]{proc_attr_gageids}
}
