% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{chck_need_vars_fmt}
\alias{chck_need_vars_fmt}
\title{Check the format of need_vars object and correct if it doesn't match}
\usage{
chck_need_vars_fmt(need_vars)
}
\arguments{
\item{need_vars}{a list whose names must follow "{shortstring}_vars", with
list contents containing actual variable names corresponding to a dataset}
}
\description{
Check the format of need_vars object and correct if it doesn't match
}
\seealso{
[proc.attr.hydfab]{proc_attr_mlti_wrap}
}
