% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_hydfab.R
\name{custom_hf_id}
\alias{custom_hf_id}
\title{Build a custom hydrofabric id that is unique to place}
\usage{
custom_hf_id(df, col_vpu = "vpu", col_id = "id")
}
\arguments{
\item{df}{The dataframe for a specific location corresponding to the}

\item{col_vpu}{The column in df representing the vpu}

\item{col_id}{The column in df representing the hydrofabric id
hydrofabric 'network'}
}
\description{
Build a custom hydrofabric id that is unique to place
}
