% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{gen_ds_gpkg}
\alias{gen_ds_gpkg}
\title{Generate dataset geopackage}
\usage{
gen_ds_gpkg(dir_db_gpkg, path_save_gpkg, gage_ids, epsg = 4326)
}
\arguments{
\item{dir_db_gpkg}{The directory containing the .gpkg database}

\item{path_save_gpkg}{The path to the dataset-sepecific .gpkg}

\item{gage_ids}{array of gage_id identifiers corresponding to the 'gage_id'
column in the geopackage files}

\item{epsg}{The EPSG code to use for the CRS; Default 4326, same as nhdplus.}
}
\description{
This implements a geopackage database containing all locations
ever acquired using RaFTS and copies the relevant gage_id locations over
to the dataset's directory. This is a .gpkg file read/write function and
doesn't return any objects.
}
\details{
Further processing inside proc_attr_gageids
will retrieve any missing locations using NLDI queries, which can take a
while! The geopackage database reduces the total NLDI queries needed for
any dataset.
}
\seealso{
\link[proc.attr.hydfab]{grab_attrs_datasets_rafts_wrap}

\link[proc.attr.hydfab]{update_gpkg_db}
}
