% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_hydfab.R
\name{parse_hfab_oconus_config}
\alias{parse_hfab_oconus_config}
\title{Parse the OCONUS hydrofabric metadata}
\usage{
parse_hfab_oconus_config(path_oconus_config)
}
\arguments{
\item{path_oconus_config}{filepath to the yaml configuration file}
}
\description{
Parse the OCONUS hydrofabric metadata
}
\details{
Reads the hydrofabric OCONUS mappings of domain-id in postal_code
format and the expected CRS of each hydrofabric file as of v2.2
}
