% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{proc_attr_gageids}
\alias{proc_attr_gageids}
\title{Process catchment attributes based on vector of gage ids.}
\usage{
proc_attr_gageids(
  gage_ids,
  featureSource,
  featureID,
  Retr_Params,
  lyrs = "network",
  overwrite = FALSE
)
}
\arguments{
\item{gage_ids}{array of gage_id values to be queried for catchment attributes}

\item{featureSource}{The \code{\link[nhdplusTools]{get_nldi_features}}feature featureSource,
e.g. 'nwissite'}

\item{featureID}{a glue-configured conversion of gage_id into a recognized
featureID for \code{\link[nhdplusTools]{get_nldi_features}}. E.g. if gage_id
represents exactly what the nldi_feature$featureID should be, then
featureID="{gage_id}". In other instances, conversions may be necessary,
e.g. featureID="USGS-{gage_id}". When defining featureID, it's expected
that the term 'gage_id' is used as a variable in glue syntax to create featureID}

\item{Retr_Params}{list. List of list structure with parameters/paths
needed to acquire variables of interest. List objects include the following:
\itemize{
\item \code{paths} list of directories or paths used to acquire and save data These include the following:
\item \code{paths$dir_db_hydfab} the local path to where hydrofabric data are saved
\item \code{path$dir_db_attrs} local path for saving catchment attributes as parquet files
\item \code{path$s3_path_hydatl} the s3 location where hydroatlas data exist
\item \code{path$dir_std_base} the location of user_data_std containing dataset that were standardized by \pkg{fs_proc}.
\item \code{datasets} character vector. A list of datasets of interest inside \code{paths$dir_std_base}. Not used in \code{proc_attr_gageids}
}}

\item{lyrs}{character. The layer names of interest from the hydrofabric gpkg. Default 'network'}

\item{overwrite}{boolean. Should the hydrofabric cloud data acquisition be redone and overwrite any local files? Default FALSE.}
}
\description{
Prepares inputs for main processing step. Iterates over each location
for grabbing catchment attribute data corresponding to the gage_id
location. Acquires user-requested variables from multiple catchment
attribute sources. Calls \code{\link{proc_attr_wrap}} which writes all
acquired variables to a parquet file as a standard data.table format.
Returns a data.table of all data returned from \code{nhdplusTools::get_nldi_feature}
that corresponded to the gage_ids
}
