% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{proc_attr_mlti_wrap}
\alias{proc_attr_mlti_wrap}
\title{Wrapper to retrieve variables from multiple comids when processing
attributes. Returns all attribute data for all comid locations}
\usage{
proc_attr_mlti_wrap(
  comids,
  Retr_Params,
  lyrs = "network",
  overwrite = FALSE,
  filter_vars = TRUE
)
}
\arguments{
\item{comids}{list of character. The common identifier USGS location codes/hf_uid/etc. for surface water features.}

\item{Retr_Params}{list. List of list structure with parameters/paths needed to acquire variables of interest}

\item{lyrs}{character. The layer names of interest from the hydrofabric gpkg. Default 'network'}

\item{overwrite}{boolean. Should the hydrofabric cloud data acquisition be redone and overwrite any local files? Default FALSE.}

\item{filter_vars}{boolean. Should the data be filtered down to the variables provided in \verb{Retr_Params$vars'?}}
}
\description{
Identifies a comid location using the hydrofabric and then
acquires user-requested variables from multiple sources. Writes all
acquired variables to a parquet file as a standard data.table format.
Re-processing runs only download data that have not yet been acquired.
}
\details{
Function returns & writes a data.table of all these fields:
featureID - e.g. USGS common identifier (default)
featureSource - e.g. "COMID" (default)
data_source - where the data came from (e.g. 'usgs_nhdplus__v2','hydroatlas__v1')
dl_timestamp - timestamp of when data were downloaded
attribute - the variable identifier used in a particular dataset
value - the value of the identifier
Note a refactor needed with \link[proc.attr.hydfab]{proc_attr_exst_wrap} if
we want to be more efficient with retrieving only the needed variables as
defined in Retr_Params$vars
}
\seealso{
\link[proc.attr.hydfab]{proc_attr_gageids}

\link[proc.attr.hydfab]{proc_attr_exst_wrap} Generates the need_vars
}
\author{
Guy Litt \email{guy.litt@noaa.gov}
}
