% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{proc_attr_wrap}
\alias{proc_attr_wrap}
\title{Wrapper to retrieve variables when processing attributes}
\usage{
proc_attr_wrap(comid, Retr_Params, lyrs = "network", overwrite = FALSE)
}
\arguments{
\item{comid}{character. The common identifier USGS location code for a surface water feature.}

\item{Retr_Params}{list. List of list structure with parameters/paths needed to acquire variables of interest}

\item{lyrs}{character. The layer names of interest from the hydrofabric gpkg. Default 'network'}

\item{overwrite}{boolean. Should the hydrofabric cloud data acquisition be redone and overwrite any local files? Default FALSE.}
}
\description{
Identifies a comid location using the hydrofabric and then
processes user-requested variables from multiple sources
}
