% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{retr_attr_hydatl}
\alias{retr_attr_hydatl}
\title{Retrieve HydroATLAS variables}
\usage{
retr_attr_hydatl(
  hf_ids,
  path_ha,
  ha_vars,
  hf_id_col = c("hf_uid", "hf_id")[2],
  s3_ha = "s3://lynker-spatial/tabular-resources/hydroATLAS/hydroatlas_vars.parquet"
)
}
\arguments{
\item{hf_ids}{character or numeric, vector or atomic. The hydrofabric ids}

\item{path_ha}{character. full path to the local parquet or s3 bucket's
parquet holding the HydroATLAS data as formatted for the hydrofabric.}

\item{ha_vars}{list of characters. The variables of interest in the HydroATLAS v1}

\item{hf_id_col}{Custom column name for hydrofabric id unique identifier}

\item{s3_ha}{character. The s3 path containing original
HydroATLAS-hydrofabric dataset.}
}
\description{
retrieves hydrofabric variables from s3 bucket or a local file
}
\seealso{
\link[proc.attr.hydfab]{retr_attr_hydatl_wrap}
}
