% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{retr_attr_hydatl_wrap}
\alias{retr_attr_hydatl_wrap}
\title{Process HydroATLAS attributes wrapper}
\usage{
retr_attr_hydatl_wrap(
  hf_ids,
  paths_ha,
  ha_vars,
  hf_id_cols = c("hf_uid", "hf_id", "divide_id", "id"),
  colname_featID = "hf_uid"
)
}
\arguments{
\item{paths_ha}{The local filepaths/s3 paths containing the HydroATLAS
data downscaled to hydrofabric divides. For more details, refer to
\href{https://www.lynker-spatial.com/data/tabular/hydroATLAS/}{hydrofabric data portal}}

\item{ha_vars}{They HydroATLAS variables of interest}

\item{hf_id_cols}{Possible hydrofabric id columns, \emph{listed in priority}, meaning
the \code{'hf_uid'} is the most important column to find, and if that is not
present, move on to the \code{'hf_id'}, then \code{'divide_id'} and finally to \code{'id'}}

\item{colname_featID}{The standard column name for location id in the output datatable}
}
\description{
Finds the attributes from a variety of HydroATLAS files, and
allows different location identifiers. Specifically, CONUS hydrofabric
locations use the comid in the 'id' column, whereas the oCONUS hydrofabric
locations may be a custom identifier hf_uid to distinguish between different
vpus such as AK and PRVI.
}
\details{
The hf_ids may contain 1) a COMID, corresponding to CONUS, &/or
2) a hydrofabric id for OCONUS
}
\seealso{
\link[proc.attr.hydfab]{std_attr_data_fmt} - uses the \code{hf_id_cols} default argument values using formals

\link[proc.attr.hydfab]{retr_attr_hydatl}

\link[proc.attr.hydfab]{custom_hf_id} for a custom hydrofabric ID with vpu in the identifier
}
