% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{retr_attr_hydatl_wrap}
\alias{retr_attr_hydatl_wrap}
\title{Process HydroATLAS attributes wrapper}
\usage{
retr_attr_hydatl_wrap(
  hf_ids,
  paths_ha,
  ha_vars,
  hf_id_cols = c("hf_uid", "hf_id", "id")
)
}
\arguments{
\item{hf_ids}{character or numeric, vector or atomic. The hydrofabric ids}

\item{ha_vars}{list of characters. The variables of interest in the HydroATLAS v1}

\item{hf_id_cols}{Possible hydrofabric id columns, \emph{listed in priority}, meaning
the \code{'hf_uid'} is the most important column to find, and if that is not
present, move on to the \code{'hf_id'} and finally to \code{'id'}}

\item{hf_id_col}{Custom column name for hydrofabric id unique identifier}

\item{s3_ha}{character. The s3 path containing original
HydroATLAS-hydrofabric dataset.}

\item{path_fs_proc}{Path used for the standardized dataset created using
\code{\link{fs_proc.proc_eval_metrics.proc_col_schema}}}

\item{dir_std_base}{The standardized base directory}

\item{ds}{The dataset name (folder inside \code{dir_std_base})}

\item{sf_comid}{sf class data.frame of comid/gage_id/geometry mappings}

\item{epsg}{The EPSG code to use for the CRS; nhdplus defaults 4326}

\item{usgs_vars}{list class. The standardized names of NHDplus variables.}

\item{dir_db_hydfab}{character class. Local directory path for storing hydrofabric data}

\item{custom_name}{character class. A custom name to insert into hydrofabric file. Default \code{glue("{lyrs}_")}}

\item{fileext}{character class. file extension of hydrofabric file. Default 'gpkg'}

\item{hf_cat_sel}{boolean. TRUE for a total catchment characterization specific to a single comid, FALSE (or anything else) for all subcatchments}

\item{hf_version}{character class. The hydrofabric version. When NULL, defaults to same as \code{hfsubsetR::get_subset()}}

\item{type}{hydrofabric type. When NULL, defaults to same as \code{hfsubsetR::get_subset()}, likely 'nextgen'}

\item{domain}{hydrofabric domain. When NULL, defaults to same as \code{hfsubsetR::get_subset()}, likely 'conus'}

\item{bucket_conn}{TODO add cloud conn details in case data stored in s3}

\item{attr_data}{list of data.frame of attribute data}

\item{need_vars}{The needed attributes that haven't been acquired yet}

\item{path_ha}{character, the filepath to hydroatlas data.}

\item{dt_cmbo}{The standardized data.table of attributes}

\item{path_attrs}{parquet filepath for attribute data}

\item{distinct_cols}{The column names in dt_new_dat that must be distinct}

\item{comids}{list of character. The common identifier USGS location codes for surface water features.}

\item{dt_all}{Dataframe/datatable of all locations and attributes}

\item{attr_vars}{List of the data source and expected attributes
(e.g. list('usgs_vars' = c("TOT_BFI","TOT_TWI")) from Retr_Params$vars)}

\item{comid}{character. The common identifier USGS location code for a surface water feature.}

\item{hfab_retr}{boolean. Should the hydrofabric geopackage data be retrieved? Default FALSE.}

\item{gage_ids}{array of gage_id values to be queried for catchment attributes}

\item{featureSource}{The \code{nhdplusTools::get_nldi_feature} feature
featureSource, e.g. 'nwissite'}

\item{featureID}{a glue-configured conversion of gage_id into a recognized
featureID for \code{nhdplusTools::get_nldi_feature}. E.g. if gage_id
represents exactly what the nldi_feature$featureID should be, then
featureID="{gage_id}". In other instances, conversions may be necessary,
e.g. featureID="USGS-{gage_id}". When defining featureID, it's expected
that the term 'gage_id' is used as a variable in glue syntax to create featureID}

\item{path_save_gpkg}{character. The filepath where the geopackage of comid/gage_id/geometry mappings get saved}

\item{loc_id_filepath}{csv filepath or dataset filepath/directory.}

\item{loc_id}{The column name of the identifier column}

\item{fmt}{The format passed to arrow::open_dataset() in the non-csv case.
Default 'csv'. May also be 'parquet', 'arrow', 'feather', 'zarr', etc.}

\item{dir_dataset}{directory path to the dataset}

\item{ds_filenames}{a matching string specific to dataset(s) of interest
inside \code{dir_dataset}}

\item{Retr_Params}{list of parameters built for grabbing catchment attribute data. List objects include the following:
\itemize{
\item \code{paths} list of directories or paths used to acquire and save data These include the following:
\item \code{paths$dir_db_hydfab} the local path to where hydrofabric data are saved
\item \code{path$dir_db_attrs} local path for saving catchment attributes as parquet files
\item \code{path$s3_path_hydatl} the s3 location where HydroATLAS data exist
\item \code{path$dir_std_base} the location of user_data_std containing dataset that were standardized by \pkg{fs_proc}.
\item \code{datasets} character vector. A list of datasets of interest inside \code{paths$dir_std_base}. If 'all' is specified, then all datasets in the directory are processed.
\item \code{ds_type} character. The identifier to use in filename when writing attribute location metadata retrieved from nhdplusTools::get_nldi_feature()
}}

\item{overwrite}{boolean default FALSE. Should hydrofabric data be overwritten?}

\item{lyrs}{default "network" the hydrofabric layers of interest.
Only 'network' is needed for attribute grabbing.}

\item{path_save_gpkg_cstm}{A custom path to save the geopackage mapping
comid/gage_id/coords. Default NULL means that the path is automated based
on the standardized input dataset location \link[proc.attr.hydfab]{std_path_retr_gpkg}}

\item{dt_site_feat}{data.table or data.frame of NLDI site features
retrieved using nhdplusTools::get_nldi_feature() and organized
using proc.attr.hydfab::proc_attr_gageids}

\item{path_meta}{the filepath for writing nldi metadata. May be parquet or csv file.}

\item{vars_ls}{A named list from Retr_Params$vars in the standardized format}

\item{attr_cfg_path}{a path to a .yaml configuration file specifying which
attributes a user is requesting}

\item{vars}{a list specifying which attributes a user is requesting, in lieu
of a list coming from a .yaml configuration file}

\item{hfab_config}{The hydrofabric-specific section from the config file, hydfab_config. list.}

\item{reqd_hfab}{The non-optional item names in the hydrofabric config file}

\item{dir_db_attrs}{The directory to the attribute database}

\item{path_missing_attrs}{The full path to file created by \link[proc.attr.hydfab]{std_path_miss_tfrm}}

\item{df_miss}{The dataframe if interested in writing. Default NULL for reading.}

\item{read}{Boolean. Should the missing data be read? Default TRUE}

\item{path_attr_config}{The file path to the attribute config file}
}
\value{
data.table of comid and geometric point in epsg 4326

List with default arguments populated corresponding to hfsubetR::get_subset(),
}
\description{
retrieves hydrofabric variables from s3 bucket or a local file

Given the standardized base directory and the dataset name,
generate the standardized geopackage

Reads in a geopackage whose filepath is
standardized by \link[proc.attr.hydfab]{std_path_retr_gpkg}
and created by \link[proc.attr.hydfab]{fs_retr_nhdp_comids_geom_wrap}

Try reading geopackage file, if it doesn't exist or is missing
locations based on the gage_ids of interest, grab them & update the
geopackage file

Removes the duplicate item corresponding to the most NA values
in a row, but only for duplicated gage_id values. This is just-in-case
a secondary attempt at grabbing a comid was successful.

Checks to see if a local dataset exists. If not, retrieve from lynker-spatial s3 bucket

Retrieves what attribute data already exists in a data storage
path for a given comid and identifies missing attributes.
Returns list of
\itemize{
\item dt_all: a data.table of existing comid data,
\item need_vars: a list of datasource ids containing a list of variable
names that will be downloaded.
}

Identifies a comid location using the hydrofabric and then
acquires user-requested variables from multiple sources. Writes all
acquired variables to a parquet file as a standard data.table format.
Re-processing runs only download data that have not yet been acquired.

DEPRECATED. Use \link{proc_attr_mlti_wrap} instead.
Identifies a single comid location using the hydrofabric and then
acquires user-requested variables from multiple sources. Writes all
acquired variables to a parquet file as a standard data.table format.
Re-processing runs only download data that have not yet been acquired.

Uses a sub-directory in the dir_db_attrs to place data

Prepares inputs for main processing step. Iterates over each location
for grabbing catchment attribute data corresponding to the gage_id
location. Acquires user-requested variables from multiple catchment
attribute sources. Calls \link{proc_attr_wrap} which writes all
acquired variables to a parquet file as a standard data.table format.
Returns a data.table of all data returned from \code{nhdplusTools::get_nldi_feature}
that corresponded to the gage_ids.

Reads directly from a csv or arrow-compatible dataset.
Returns the dataset's column identifer renamed as 'gage_id' in a tibble

Reads output generated using \pkg{fs_proc} python package and
selects the gage_id location identifier(s) and the
featureSource & featureID that correspond to the gage_id

Wrapper function that acquires catchment attribute data from
formulation-selector processed input generated via \pkg{fs_proc} package
Returns list of data.table object of nhdplusTools::get_nldi_feature()
for all gage_ids

A short description...

Given a list of variable categories, each containing vectors
of variable names, check the following:
\enumerate{
\item the variable category is a recognized category name (e.g. 'usgs_vars')
\item the variable names inside the category name are actual variable names
that can be used to retrieve attributes (e.g. 'TOT_TWI' as an nhdplus attribute)
}

Sees if the attributes requested by a user matches with the
attributes supported by the package, listed in the attribute menu.
Returns list of
\itemize{
\item missing_vars: a list of the requested variables that were not found
in the attribute menu as specified.
}

If an argument provided in the config file is NULL, first look
for default param value from the \code{proc.attr_hydfab::proc_attr_hf}.
If that is NULL, then look for default value from the
\code{hfsubsetR::get_subset()} args if that param exists there. Otherwise,
uses default param value in \code{proc.attr_hydfab::proc_attr_wrap}.
}
\details{
The hf_ids may contain 1) a COMID, corresponding to CONUS, &/or
2) a hydrofabric id for OCONUS

The python companion function is
. \code{fs_algo.fs_algo_train_eval._std_fs_proc_ds_companion_gpkg_path}

These same steps happen inside \code{grab_attrs_datasets_fs_wrap}

The renaming may be a temporary solution just-in case file
originated in python fs_algo

This can change the exact 1:1 match of gage_ids, which is expected
in \link[proc.attr.hydfab]{proc_attr_gageids}. May set epsg to NULL

Checks to see if data already exists. If so, read it in. Then
merges new data with existing data and remove any duplicates

Function returns & writes a data.table of all these fields:
featureID - e.g. USGS common identifier (default)
featureSource - e.g. "COMID" (default)
data_source - where the data came from (e.g. 'usgs_nhdplus__v2','hydroatlas__v1')
dl_timestamp - timestamp of when data were downloaded
attribute - the variable identifier used in a particular dataset
value - the value of the identifier

Writes to file the missing comid-attribute pairings after
first updating the existing known missing data

Function returns & writes a data.table of all these fields:
featureID - e.g. USGS common identifier (default)
featureSource - e.g. "COMID" (default)
data_source - where the data came from (e.g. 'usgs_nhdplus__v2','hydroatlas__v1')
dl_timestamp - timestamp of when data were downloaded
attribute - the variable identifier used in a particular dataset
value - the value of the identifier

The gage_id-comid mappings are saved to file to avoid exceeding
the NLDI database connection rate limit

Intendended for a single dataset, aka input file directory
containing data. For multiple datasets, refer to dataset looping inside
\link[proc.attr.hydfab]{grab_attrs_datasets_fs_wrap}

Returns a list of the following objects:
gage_ids: array of gage_id values
featureSource: The type of nhdplus feature source corresponding to gage_id
featureID: The method of converting gage_id into a standardized featureSource's featureID

Runs two proc.attr.hydfab functions:
\link[proc.attr.hydfab]{proc_attr_read_gage_ids_fs} - retrieves the gage_ids generated by \pkg{fs_proc}
\link[proc.attr.hydfab]{proc_attr_gageids} - retrieves the attributes for all provided gage_ids

Use fs_attrs_miss_mlti_wrap instead.
Given missing comid-attribute pairings previously identified
from fs_tfrm_attrs.py, and generated as a file by python function
\code{fs_algo.tfrm_attr.write_missing_attrs}

Given missing comid-attribute pairings previously identified
from fs_tfrm_attrs.py, and generated as a file by python function
\code{fs_algo.tfrm_attr.write_missing_attrs}
}
\note{
2025-03-07 This needs a deeper refactoring
}
\seealso{
\link[proc.attr.hydfab]{retr_attr_hydatl}

\link[proc.attr.hydfab]{retr_attr_hydatl_wrap}

\link[proc.attr.hydfab]{std_path_retr_gpkg}

\code{fs_algo.fs_algo_train_eval._std_fs_proc_ds_companion_gpkg_path}

\link[proc.attr.hydfab]{read_fs_retr_gpkg}

\link[proc.attr.hydfab]{std_dir_dataset}

\link[proc.attr.hydfab]{fs_retr_nhdp_comids_geom_wrap}

\link[proc.attr.hydfab]{std_path_retr_gpkg}

\link[proc.attr.hydfab]{fs_retr_nhdp_comids_geom_wrap}

\link[proc.attr.hydfab]{proc_attr_read_gage_ids_fs}

\link[proc.attr.hydfab]{fs_retr_nhdp_comids_geom}

\code{fs_algo.fs_algo_train_eval.fs_retr_nhdp_comids_geom_wrap}

\link[proc.attr.hydfab]{fs_retr_nhdp_comids_geom_wrap}

\link[proc.attr.hydfab]{proc_attr_gageids}

\link[proc.attr.hydfab]{proc_attr_read_gage_ids_fs}

\link[proc.attr.hydfab]{fs_retr_nhdp_comids_geom_wrap}

\code{fs_algo.fs_algo_train_eval.fs_retr_nhdp_comids_geom}

\link[nhdplusTools]{get_characteristics_metadata}

\link[proc.attr.hydfab]{proc_attr_wrap}

\link[proc.attr.hydfab]{retr_attr_new}

\link[proc.attr.hydfab]{proc_attr_wrap}

\link[proc.attr.hydfab]{proc_attr_mlti_wrap}

\link[proc.attr.hydfab]{proc_attr_wrap}

\code{fs_algo.fs_algo_train_eval.fs_read_attr_comid()} python function
that reads these files

\link[proc.attr.hydfab]{retrieve_attr_exst} for retrieving existing attributes

\link[proc.attr.hydfab]{std_attr_data_fmt}

\link[proc.attr.hydfab]{std_path_attrs}

\link[proc.attr.hydfab]{proc_attrs_gageids}

\link[proc.attr.hydfab]{proc_attr_mlti_wrap}

\link[proc.attr.hydfab]{retr_attr_new}

\link[proc.attr.hydfab]{proc_attrs_gageids}

\link[proc.attr.hydfab]{proc_attr_mlti_wrap}

\link[proc.attr.hydfab]{grab_attrs_datasets_fs_wrap} loops over multiple datasets in calling this gage_id retrieval

\link[proc.attr.hydfab]{proc_attr_mlti_wrap} Acquire attributes

\link[proc.attr.hydfab]{proc_attr_read_gage_ids_fs}

\link[proc.attr.hydfab]{proc_attr_wrap}

\link[nhdplusTools]{get_nldi_feature}

\link[proc.attr.hydfab]{proc_attr_gageids}

\code{fs_algo.tfrm_attrs.std_path_miss_tfrm} python package

\code{fs_algo.tfrm_attr.write_missing_attrs} python

\link[proc.attr.hydfab]{fs_attrs_miss_mlti_wrap}

\link[proc.attr.hydfab]{fs_attrs_miss_mlti_wrap}

\code{fs_algo.tfrm_attr.write_missing_attrs} python

\link{fs_attrs_miss.R} Rscript that calls this wrapper
}
\author{
Guy Litt \email{guy.litt@noaa.gov}

Lauren Bolotin \email{lauren.bolotin@noaa.gov }
}
