% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{retr_attr_hydatl_wrap}
\alias{retr_attr_hydatl_wrap}
\title{Process HydroATLAS attributes wrapper}
\usage{
retr_attr_hydatl_wrap(
  hf_ids,
  paths_ha,
  ha_vars,
  hf_id_cols = c("hf_uid", "hf_id", "id")
)
}
\arguments{
\item{hf_id_cols}{Possible hydrofabric id columns, \emph{listed in priority}, meaning
the \code{'hf_uid'} is the most important column to find, and if that is not
present, move on to the \code{'hf_id'} and finally to \code{'id'}}
}
\description{
Process HydroATLAS attributes wrapper
}
\details{
The hf_ids may contain 1) a COMID, corresponding to CONUS, &/or
2) a hydrofabric id for OCONUS
}
\seealso{
\link[proc.attr.hydfab]{retr_attr_hydatl}
}
