% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{retr_attr_new}
\alias{retr_attr_new}
\title{Retrieve new attributes that haven't been acquired yet}
\usage{
retr_attr_new(locids, need_vars, paths_ha)
}
\arguments{
\item{locids}{The list of of the comid or hydrofabric unique location identifier}

\item{need_vars}{The needed attributes that haven't been acquired yet}

\item{paths_ha}{vector, the filepath(s) to HydroATLAS data downscaled to Hydrofabric, in tabular form}

\item{Retr_Params}{list. List of list structure with parameters/paths needed to acquire variables of interest}
}
\description{
Retrieve new attributes that haven't been acquired yet
}
\seealso{
\link[proc.attr.hydfab]{proc_attr_wrap}

\link[proc.attr.hydfab]{proc_attr_mlti_wrap}
}
