% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{retr_attr_new}
\alias{retr_attr_new}
\title{Retrieve new attributes that haven't been acquired yet}
\usage{
retr_attr_new(comids, need_vars, Retr_Params)
}
\arguments{
\item{comids}{The list of of the comid identifier}

\item{need_vars}{The needed attributes that haven't been acquired yet}

\item{Retr_Params}{list. List of list structure with parameters/paths needed to acquire variables of interest}
}
\description{
Retrieve new attributes that haven't been acquired yet
}
\seealso{
\link{proc_attr_wrap}

\link{proc_attr_mlti_wrap}
}
