% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{retr_comids}
\alias{retr_comids}
\title{Retrieve comids based on provided gage_ids and expected NLDI format}
\usage{
retr_comids(gage_ids, featureSource, featureID, dir_db_attrs)
}
\arguments{
\item{gage_ids}{array of gage_id values to be queried for catchment attributes}

\item{featureSource}{The \link[nhdplusTools:get_nldi_feature]{nhdplusTools::get_nldi_feature}feature featureSource,
e.g. 'nwissite'}

\item{featureID}{a glue-configured conversion of gage_id into a recognized
featureID for \link[nhdplusTools:get_nldi_feature]{nhdplusTools::get_nldi_feature}. E.g. if gage_id
represents exactly what the nldi_feature$featureID should be, then
featureID="{gage_id}". In other instances, conversions may be necessary,
e.g. featureID="USGS-{gage_id}". When defining featureID, it's expected
that the term 'gage_id' is used as a variable in glue syntax to create featureID}
}
\description{
Retrieve comids based on provided gage_ids and expected NLDI format
}
\details{
The gage_id-comid mappings are saved to file to avoid exceeding
the NLDI database connection rate limit
}
