% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_hydfab.R
\name{retr_comids_coords}
\alias{retr_comids_coords}
\title{Iterate over each coordinate in a dataframe to retrieve comid}
\usage{
retr_comids_coords(
  df,
  col_lat = "latitude",
  col_lon = "longitude",
  col_crs = NULL,
  crs = 4326
)
}
\arguments{
\item{df}{Dataframe containing lat/lon columns}

\item{col_lat}{The column name for the latitude data}

\item{col_lon}{The column name for the longitude data}

\item{col_crs}{Default NULL, column name for the CRS. If NULL, \code{crs} arg
will be used.}

\item{crs}{The CRS to use when \code{col_crs} is NULL. Default 4326.}
}
\description{
Iterate over each coordinate in a dataframe to retrieve comid
}
