% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_hydfab.R
\name{retr_hfab_id_coords}
\alias{retr_hfab_id_coords}
\title{Retrieve hydrofabric IDs via coordinates}
\usage{
retr_hfab_id_coords(path_gpkg, ntwk, epsg_domn, lon, lat, epsg_coords = 4326)
}
\arguments{
\item{path_gpkg}{Path to the hydrofabric geopackage file}

\item{ntwk}{The network hydrofabric layer}

\item{epsg_domn}{The EPSG of the hydrofabric domain of interest}

\item{lon}{Longitude}

\item{lat}{Latitude}

\item{epsg_coords}{The EPSG code corresponding to \code{lat} and \code{lon}. Default 4326}
}
\description{
Retrieve hydrofabric IDs via coordinates
}
\details{
Assumes the hydrofabric domain (e.g. AK, HI) is already determined
via the path_gpkg
}
