% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_hydfab.R
\name{retr_hfab_id_wrap}
\alias{retr_hfab_id_wrap}
\title{Retrieve hydrofabric IDs wrapper}
\usage{
retr_hfab_id_wrap(
  dt_need_hf,
  path_oconus_hfab_config,
  col_usgsId = "usgsId",
  col_lon = "longitude",
  col_lat = "latitude",
  epsg_coords = 4326
)
}
\arguments{
\item{dt_need_hf}{data.table of needed}

\item{col_usgsId}{column name inside \code{dt_need_hf} for the USGS gage ID}

\item{col_lon}{column name inside \code{dt_need_hf} for longitude value}

\item{col_lat}{column name inside \code{dt_need_hf} for latitude value}

\item{epsg_coords}{The CRS for the lat/lon data inside \code{dt_need_hf}}

\item{path_oconus_config}{File path to yaml config mapping of OCONUS
hydrofabric locations, postal codes, and expected CRS geopackage filepaths}
}
\description{
Retrieve hydrofabric IDs wrapper
}
\details{
Intended for situations when comids unavailable, generally as OCONUS
}
\seealso{
prep_oconus_hydroatlas.R also createas the hf_uid using custom_hf_id()
}

