% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{std_attr_data_fmt}
\alias{std_attr_data_fmt}
\title{Standardize the catchment attribute data to read/write in parquet files}
\usage{
std_attr_data_fmt(attr_data)
}
\arguments{
\item{attr_data}{list of data.frame of attribute data}
}
\description{
Standardize the catchment attribute data to read/write in parquet files
}
\seealso{
\link{retr_attr_new}
}
