% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafts_utils.R
\name{std_dir_logs}
\alias{std_dir_logs}
\title{Define the standard directory used for storing log files}
\usage{
std_dir_logs(dir_db_attrs)
}
\arguments{
\item{dir_db_attrs}{The path where attribute data are stored}
}
\description{
Define the standard directory used for storing log files
}
\seealso{
fs_prep.std_dir_log (python w/ different arg)
}
