% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{std_path_retr_gpkg}
\alias{std_path_retr_gpkg}
\title{Create the standardized gpkg path for coordinate data & id mapping
. corresponding to the standardized input data}
\usage{
std_path_retr_gpkg(path_fs_proc)
}
\arguments{
\item{path_fs_proc}{Path used for the standardized dataset created using
\code{\link{fs_proc.proc_eval_metrics.proc_col_schema}}}
}
\description{
Create the standardized gpkg path for coordinate data & id mapping
. corresponding to the standardized input data
}
\details{
The python companion function is
. \code{fs_algo.fs_algo_train_eval._std_fs_proc_ds_companion_gpkg_path}
}
\seealso{
\code{fs_algo.fs_algo_train_eval._std_fs_proc_ds_companion_gpkg_path}

\link[proc.attr.hydfab]{read_fs_retr_gpkg}

\link[proc.attr.hydfab]{std_dir_dataset}
}
