% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{std_write_geom_map_gpkg}
\alias{std_write_geom_map_gpkg}
\title{Remove duplicates and write comid-geometry mappings to file}
\usage{
std_write_geom_map_gpkg(sf_comid, path_save_gpkg)
}
\arguments{
\item{sf_comid}{sf class data.frame of comid/gage_id/geometry mappings}
}
\description{
Removes the duplicate item corresponding to the most NA values
in a row, but only for duplicated gage_id values. This is just-in-case
a secondary attempt at grabbing a comid was successful.
}
\details{
This can change the exact 1:1 match of gage_ids, which is expected
in \link[proc.attr.hydfab]{proc_attr_gageids}
}
\seealso{
\link[proc.attr.hydfab]{fs_retr_nhdp_comids_geom_wrap}

\link[proc.attr.hydfab]{proc_attr_gageids}
}
