% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_attr_grabber.R
\name{update_gpkg_db}
\alias{update_gpkg_db}
\title{Update the geopackage database with the dataset's gpkg}
\usage{
update_gpkg_db(dir_db_gpkg, path_save_gpkg, epsg = 4326)
}
\arguments{
\item{dir_db_gpkg}{The directory containing the .gpkg database}

\item{path_save_gpkg}{The path to the dataset-sepecific .gpkg}

\item{epsg}{The EPSG code to use for the CRS; Default 4326, same as nhdplus.}
}
\description{
This is a .gpkg file read/write function and doesn't return
any objects.
}
\details{
\code{gen_ds_gpkg} Updates the dataset gpkg in the dataset dir
using data that already exist in the gpgk database.
\code{update_gpkg_db} is performed after retrieving missing gage_id
locations, and takes any additions to the dataset gpkg and adds them
to the gpkg database.
}
\seealso{
\link[proc.attr.hydfab]{gen_ds_gpkg}

\link[proc.attr.hydfab]{grab_attrs_datasets_fs_wrap}
}
